requirements.txt
asyncpg==0.29.0
fastapi==0.89.1
psycopg2-binary==2.9.9
uvicorn==0.20.0
argon2-cffi==21.1.0
fastapi_utils==0.2.0
python-multipart==0.0.9
Jinja2==3.1.3
PyGobject==3.46.0
weasyprint==61.1
qrcode==7.4.2
httpx==0.27.0
pytest==8.0.2
sqlalchemy==1.4.52
python-dotenv==1.0.1
pytz==2024.1 